Rename to  xmtp_utils
