+++ notification-worker/src/xmtp_utils.rs
